+++ frontend/controllers/ScoreController.php
@@ -2,12 +2,16 @@
 
 namespace frontend\controllers;
 
+use ciniran\excel\SaveExcel;
+use common\models\config;
 use frontend\models\Student;
 use Yii;
 use frontend\models\Score;
 use frontend\models\ScoreSearch;
+use yii\data\Pagination;
 use yii\helpers\Url;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -37,13 +41,52 @@ class ScoreController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new ScoreSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if (!Yii::$app->user->can('indexScore')){
+            throw new ForbiddenHttpException(Yii::$app->params['perMessage']);
+        }
+        $searchCondition = Yii::$app->request->queryParams; //获得搜索参数
 
+        $searchModel = new ScoreSearch();
+
+        $dataProvider = $searchModel->search($searchCondition);
+
+        if (isset($searchCondition["ScoreSearch"]["isExport"]) && $searchCondition["ScoreSearch"]["isExport"] == 1){
+            if (!Yii::$app->user->can('exportScore')){
+                throw new ForbiddenHttpException(Yii::$app->params['perMessage']);
+            }
+            //导出不适应分页，把默认的20条改的很大
+            $dataProvider->setPagination(new Pagination([
+                'defaultPageSize' => 1000000,
+                'pageSizeLimit' => [1, 1000000]
+            ]));
+
+            $models = $dataProvider->getModels();
+
+            $searchModel->dealExportData($models); //处理模型数据
+
+            $fields = $searchModel->type == 1 ? ['cand_id','name','test_name','grade','banji','chinese','math','english',
+                                                'physics','chemistry','biology','total','class_rank','school_rank', 'type'] : ['cand_id','name','test_name','grade','banji','chinese','math','english',
+                                                'politics','history','geography','total','class_rank','school_rank', 'type'
+                                                ];
+
+            $fileName = $searchCondition['ScoreSearch']['test_name'].'学生成绩表'.time();
+
+            $excel = new SaveExcel([
+
+                'models' => $models,
+
+                'fields' => $fields,
+
+                'fileName' => $fileName
+            ]);
+
+            $excel->modelsToExcel();
+        }
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
             'type' => $searchModel->type,
+            'searchCondition' => $searchCondition,
 
         ]);
     }
@@ -56,6 +99,9 @@ class ScoreController extends Controller
      */
     public function actionView($id)
     {
+        if (!Yii::$app->user->can('viewScore')){
+            return $this->renderAjax('/site/error',['name'=>'权限验证不通过','message'=>Yii::$app->params['perMessage']]);
+        }
         $model = $this->findModel($id);
 
         $img = (Student::find()->where(['student_id'=>$model->student_id])->one())->pic;
@@ -99,6 +145,9 @@ class ScoreController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->can('updateScore')){
+            return $this->renderAjax('/site/error',['name'=>'权限验证不通过','message'=>Yii::$app->params['perMessage']]);
+        }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -123,6 +172,9 @@ class ScoreController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->can('deleteScore')){
+            throw new ForbiddenHttpException(Yii::$app->params['perMessage']);
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
